ui: Select clusterVersion, clusterId, and clusterName on `layout` container
- Move clusterVersion selector to `nodes` reducer as this value is retrieved from nodes description;
- Select clusterVersion, clusterId, and clusterName on `layout` container.

Release note: None
